refactor: create script utils with base scenario for simulations
