+++ scripts/logPTokenData.ts
@@ -3,7 +3,7 @@ import { baseSepolia } from 'viem/chains';
 import dotenv from 'dotenv';
 import { PTokenAbi } from '../abis/PTokenAbi';
 import { resolve } from 'path';
-import { validateEnvKeys } from './validateEnvKeys';
+import { validateEnvKeys } from './utils/validateEnvKeys';
 
 // Load environment variables
 dotenv.config({ path: resolve(process.cwd(), '.env.local') });
