+++ scripts/utils/validateEnvKeys.ts
